fix(tcp): got server working
